100% coverage
Enable coveralls to report on this.

Coverage reports showed two test classes that weren't subclassing the
correct base class, so we weren't actually testing the C
implementation of `removeAllProxies` and `queryInnerProxy`. Fixing
this revealed a probably bug in the implementation of
`removeAllProxies`; right now that test is skipped.

Also remove the `test_suite` for 'python setup.py test' since it
usually doesn't work anymore. See #18.